04 again

+++ waitercaller.py
@@ -122,16 +122,20 @@ def account():
 @login_required
 def account_createtable():
     form = CreateTableForm(request.form)
+    flag = 0
     if form.validate():
         tableid = DB.add_table(form.tablenumber.data, current_user.get_id())
         new_url = BH.shorten_url(config.base_url + "newrequest/" + str(tableid))
-        DB.delete_request_redundancy(tableid)
         DB.update_table(tableid, new_url)
+        flag = 1 
         #tableid02 = DB.add_table_fulltest(form.tablenumber.data, current_user.get_id(),new_url)
         #DB.delete_table_fulltest(tableid02)
         
         return redirect(url_for('account'))
-    
+    if flag = 1:
+        tableid02 = DB.add_table_fulltest(form.tablenumber.data, current_user.get_id(),new_url)
+        DB.delete_request_redundancy(tableid)
+        flag = 0
     return render_template("account.html", createtableform=form, tables=DB.get_tables(current_user.get_id()))
 
 @app.route("/account/deletetable")
